Moved file association logic to it's dedicated file
